global: move from setup.py to setup.cfg
