+++ .github/workflows/pypi-publish.yml
@@ -15,7 +15,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install setuptools wheel
+          pip install setuptools wheel babel
       - name: Build package
         run: |
           python setup.py compile_catalog sdist bdist_wheel
