+++ .gitignore
@@ -30,7 +30,6 @@ var/
 *.egg
 Pipfile
 Pipfile.lock
-pyproject.toml
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
