+++ invenio_users_resources/__init__.py
@@ -9,15 +9,8 @@
 
 """Invenio module providing management APIs for users and roles/groups."""
 
-from .ext import InvenioUsersResources
-from .proxies import current_groups_service, current_user_resources, \
-    current_users_service
-from .version import __version__
+__version__ = "0.1.0"
 
 __all__ = (
     "__version__",
-    "current_user_resources",
-    "current_users_service",
-    "current_groups_service",
-    "InvenioUsersResources",
 )
