+++ setup.cfg
@@ -6,6 +6,53 @@
 # modify it under the terms of the MIT License; see LICENSE file for more
 # details.
 
+[metadata]
+name = invenio-users-resources
+version = attr: invenio_users_resources.__version__
+description = "Invenio module providing management APIs for users and roles/groups."
+long_description = file: README.rst, CHANGES.rst
+keywords = invenio users roles groups management
+license = MIT
+author = CERN/TU Wien/JRC
+author_email = info@inveniosoftware.org
+platforms = any
+url = https://github.com/inveniosoftware/invenio-users-resources
+classifiers =
+    Development Status :: 3 - Alpha
+
+[options]
+include_package_data = True
+packages = find:
+python_requires = >=3.6
+zip_safe = False
+install_requires =
+    # OAuthClient brings in Invenio-Accounts and Invenio-I18N
+    invenio-oauthclient>=2.0.0.dev0
+    invenio-records-resources>=0.19.0,<0.20
+    invenio-userprofiles>=1.2.4
+
+[options.extras_require]
+tests =
+    invenio-app>=1.3.2
+    invenio-db[postgresql,mysql,versioning]>=1.0.13,<2.0
+    pytest-invenio>=1.4.4
+    sphinx>=4.2.0,<5
+elasticsearch7 =
+    invenio-search[elasticsearch7]>=1.4.2,<2.0
+
+[options.entry_points]
+invenio_base.apps =
+    invenio_users_resources = invenio_users_resources.ext:InvenioUsersResources
+invenio_base.api_apps =
+    invenio_users_resources = invenio_users_resources.ext:InvenioUsersResources
+invenio_base.api_blueprints =
+    invenio_users = invenio_users_resources.views:create_users_resources_bp
+    invenio_groups = invenio_users_resources.views:create_groups_resources_bp
+invenio_search.mappings =
+    users = invenio_users_resources.records.mappings
+    groups = invenio_users_resources.records.mappings
+invenio_i18n.translations =
+    messages = invenio_users_resources
 
 [build_sphinx]
 source-dir = docs/
